refactor: 컴포넌트 세분화 및 재사용
