+++ src/components/CoinImg.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+interface IProps {
+  coinImg: string;
+}
+
+const CoinImg = ({ coinImg }: IProps) => {
+  return <img src={`https://static.upbit.com/logos/${coinImg}.png`} alt="" />;
+};
+
+export default CoinImg;
