+++ src/components/CoinTitle.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+interface IProps {
+  title: string;
+}
+
+const CoinTitle = ({ title }: IProps) => {
+  return <p>{title}</p>;
+};
+
+export default CoinTitle;
